% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaletteCode.R
\name{colorblind.friendly}
\alias{colorblind.friendly}
\title{Colorblind-Friendly Palette Check}
\usage{
colorblind.friendly(palette.name)
}
\arguments{
\item{name}{Name of Palette. Choices are:
\code{Austria}, \code{Cassatt}, \code{Degas}, \code{Egypt}, \code{Gauguin}, \code{Greek},
\code{Hokusai}, \code{Ingres}, \code{Isfahan1}, \code{Isfahan2}, \code{Juarez}, \code{Klimt},
\code{Manet}, \code{Monet}, \code{Moreau}, \code{Morgenstern}, \code{Nattier}, \code{NewKingdom},
\code{Pillement}, \code{Pissaro}, \code{Redon}, \code{Renoir}, \code{Robert}, \code{Stevens}, \code{Tara},
\code{Thomas}, \code{Tiepolo}, \code{Troy}, \code{VanGogh1}, \code{VanGogh2}, \code{Veronese}, and \code{Wissing}}
}
\value{
TRUE/FALSE if palette is colorblind-friendly
}
\description{
Checks whether a palette is colorblind-friendly.
}
\examples{
colorblind.friendly("Veronese")
}
