% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaletteCode.R
\docType{data}
\name{colorblind_palettes}
\alias{colorblind_palettes}
\title{Names of colorblind-friendly palettes}
\format{
An object of class \code{character} of length 18.
}
\usage{
colorblind_palettes
}
\description{
Lists all palettes that are colorblind-friendly in the package.
To be colorblind-friendly, all colors in the palettes must be distinguishable with deuteranopia, protanopia, and tritanopia.
Use \code{\link{met.brewer}} to construct palettes or \code{\link{colorblind.friendly}} to test for colorblind-friendliness.
}
\keyword{datasets}
