% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetBrewer-package.R
\docType{package}
\name{MetBrewer-package}
\alias{MetBrewer}
\alias{MetBrewer-package}
\title{MetBrewer: Color Palettes Inspired by Works at the Metropolitan Museum of Art}
\description{
Palettes Inspired by Works at the Metropolitan Museum of Art in New York. Currently contains over 50 color schemes and checks for colorblind-friendliness of palettes. Colorblind accessibility checked using the '{colorblindcheck} package by Jakub Nowosad'\url{https://nowosad.github.io/colorblindcheck/}.
}
\keyword{internal}
