% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaletteCode.R
\docType{data}
\name{colorblind_palettes}
\alias{colorblind_palettes}
\title{Names of colorblind-friendly palettes}
\format{
An object of class \code{character} of length 18.
}
\usage{
colorblind_palettes
}
\description{
Use \code{\link{met.brewer}} to construct palettes.
}
\keyword{datasets}
