% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaletteCode.R
\docType{data}
\name{MetPalettes}
\alias{MetPalettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 31.
}
\usage{
MetPalettes
}
\description{
Use \code{\link{met.brewer}} to construct palettes.
}
\keyword{datasets}
